# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Happy Yatra API
  description: |
    REST API for **Happy Yatra** — a personalized travel destination recommendation platform.

    ## Authentication
    Protected routes require a Bearer JWT token in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```
    Obtain tokens via `/api/auth/login` or `/api/auth/signup`.

    ## Token Lifecycle
    - **Access Token**: expires in **1 day**
    - **Refresh Token**: expires in **7 days**
    - Tokens are invalidated server-side via the **Token Blacklist** on logout.
  version: "1.0.0"
  contact:
    name: Happy Yatra
    url: https://happyyatra.netlify.app
servers:
  - url: https://happyatra.vercel.app/api
    description: Production (Vercel)
  - url: http://localhost:9000/api
    description: Local Development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Access Token. Obtain from /api/auth/login or /api/auth/signup."
  schemas:
...
    Destination:
      type: object
      required:
        - name
        - imageUrl
        - averageCost
        - styles
        - tags
        - activities
        - location
        - latitude
        - longitude
      properties:
        _id:
          type: string
          example: "64xyz789abc123def456ghi"
        name:
          type: string
          example: "Goa"
        imageUrl:
          type: string
          format: uri
          example: "https://example.com/goa.jpg"
        averageCost:
          type: number
          minimum: 0
          example: 15000
          description: Average cost in INR
        styles:
          type: array
          items:
            type: string
          minItems: 1
          example: ["beach", "relaxation"]
        tags:
          type: array
          items:
            type: string
          minItems: 1
          example: ["tropical", "coastal"]
        activities:
          type: array
          items:
            type: string
          minItems: 1
          example: ["surfing", "sightseeing"]
        location:
          type: string
          example: "Goa, India"
        latitude:
          type: number
          minimum: -90
          maximum: 90
          example: 15.2993
        longitude:
          type: number
          minimum: -180
          maximum: 180
          example: 74.1240
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Survey:
      type: object
      required:
        - user
        - travelStyle
        - budget
      properties:
        _id:
          type: string
        user:
          type: string
          description: MongoDB ObjectId referencing the Users collection
          example: "64abc123def456ghi789jkl"
        travelStyle:
          type: string
          example: "adventure"
        budget:
          type: number
          example: 30000
        interests:
          type: array
          items:
            type: string
          default: []
          example: ["nature", "culture"]
        activities:
          type: array
          items:
            type: string
          default: []
          example: ["hiking", "photography"]
        createdAt:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        message:
          type: string
          example: "Human-readable error message"
        error:
          type: string
          description: Technical error detail (optional)
          example: "Invalid email or password"
paths:
...
  /destinations:
    get:
      tags: [Destinations]
      summary: Get paginated destinations
      description: |
        Fetch a paginated list of travel destinations.
        Use `trending=true` to filter trending destinations.

        **Warning**: No pagination metadata (total count, totalPages) is returned in the response.
      operationId: getDestinations
      parameters:
        - name: page
          in: query
          description: Page number (1-indexed)
          schema:
            type: integer
            default: 1
            minimum: 1
          example: 1
        - name: limit
          in: query
          description: Number of results per page
          schema:
            type: integer
            default: 12
            minimum: 1
            maximum: 100
          example: 12
        - name: trending
          in: query
          description: Filter to only trending destinations
          schema:
            type: string
            enum: ["true", "false"]
          example: "true"
      responses:
        "200":
          description: Array of destinations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Destination"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: "Server Error on get Destinations"
...
  /survey:
    post:
      tags: [Survey]
      summary: Submit travel preference survey
      description: |
        Submit travel preferences tied to a user ObjectId.

        **Warning**: This endpoint is currently unauthenticated. The `user` ObjectId
        is trusted from the request body without server-side verification of identity
        or existence of the referenced user.
      operationId: submitSurvey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user, travelStyle, budget]
              properties:
                user:
                  type: string
                  description: Valid MongoDB ObjectId referencing a user
                  example: "64abc123def456ghi789jkl"
                travelStyle:
                  type: string
                  example: "adventure"
                budget:
                  type: number
                  example: 30000
                interests:
                  type: array
                  items:
                    type: string
                  default: []
                  example: ["nature", "culture"]
                activities:
                  type: array
                  items:
                    type: string
                  default: []
                  example: ["hiking", "photography"]
      responses:
        "201":
          description: Survey submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Survey submitted successfully"
        "400":
          description: Invalid user ObjectId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: "Invalid user ID"
        "500":
          description: Database error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to submit survey"

    get:
      tags: [Survey]
      summary: Get all surveys
      description: |
        Retrieve all survey submissions sorted by newest first.

        **Warning**: This endpoint is currently unauthenticated and exposes all user survey data.
        It should be restricted to admin users only.
      operationId: getSurveys
      responses:
        "200":
          description: Array of survey submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Survey"
        "500":
          description: Database error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to fetch survey data"